+++ packages/protocol/src/index.ts
@@ -18,3 +18,4 @@ export * from './fingerprint.js';
 export * from './ids.js';
 export * from './ratelimit.js';
 export * from './responses.js';
+export * from './sourcemap.js';